Split adcore io and logic into seperate files
